Updating 101-print_number.c file

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -10,7 +10,7 @@
 
 void print_number(int n)
 {
-unsigned long num = n;
+unsigned int num = n;
 
 if (n < 0)
 {
